feature/AND-BG Звезды в космосе на фоне (#23)
* Звезды в космосе на фоне

* Add cursor game

* убрал курсор
