--- packages/client/src/App.css
@@ -1,7 +1,0 @@
-.app {
-  width: 100vw;
-  height: 100vh;
-  overflow-y: auto;
-  background: var(--backgroundMain);
-  color: var(--color-light);
-}
