+++ packages/client/src/App.module.css
@@ -1,0 +1,65 @@
+.app {
+  width: 100vw;
+  height: 100vh;
+  overflow: hidden;
+  background: var(--backgroundMain);
+  color: var(--color-light);
+  position: relative;
+
+  &__star {
+    position: absolute;
+    background: radial-gradient(
+      ellipse at center,
+      rgba(177, 198, 219, 1) 2%,
+      rgba(5, 63, 118, 1) 100%
+    );
+    border-radius: 100%;
+    animation: shine infinite alternate;
+  }
+
+  &__content {
+    width: 100vw;
+    height: 100vh;
+    overflow: hidden;
+    position: relative;
+    text-align: center;
+    z-index: 2;
+  }
+}
+
+@keyframes shine {
+  0% {
+    transform: scale(1);
+    opacity: 1;
+  }
+
+  20% {
+    transform: scale(0.9);
+    opacity: 0.8;
+  }
+
+  40% {
+    transform: scale(1);
+    opacity: 0.9;
+  }
+
+  40% {
+    transform: scale(0.2);
+    opacity: 0.2;
+  }
+
+  60% {
+    transform: scale(0.2);
+    opacity: 0.1;
+  }
+
+  80% {
+    transform: scale(0.5);
+    opacity: 0.5;
+  }
+
+  100% {
+    transform: scale(0.9);
+    opacity: 0.9;
+  }
+}
