--- packages/client/src/components/ErrorSample/constants.ts
@@ -1,12 +1,0 @@
-export const CONFIG_PARAMS = {
-  amount: 200,
-  size: {
-    min: 1,
-    max: 5,
-    giant: 9,
-  },
-  duration: {
-    min: 5,
-    max: 25,
-  },
-} as const;
