+++ packages/client/src/constants/vars.ts
@@ -5,3 +5,16 @@ enum Language {
 
 export const BASE_URL = 'https://ya-praktikum.tech/api/v2';
 export const LANGUAGE = Language.eng;
+
+export const CONFIG_STARS_PARAMS = {
+  amount: 200,
+  size: {
+    min: 1,
+    max: 5,
+    giant: 9,
+  },
+  duration: {
+    min: 10,
+    max: 40,
+  },
+} as const;
