+++ packages/client/src/index.css
@@ -33,7 +33,6 @@ input {
 }
 
 .main-menu {
-  background: var(--color-bg);
   height: 100vh;
   width: 100vw;
   position: fixed;
