+++ packages/client/src/pages/Game/index.tsx
@@ -11,9 +11,8 @@ export default function GamePage() {
       game.current = new Game({ canvas: canvas.current });
       game.current?.init();
     }
-
     return () => game?.current?.unmount();
   }, [canvas]);
 
-  return <canvas className={'canvas'} ref={canvas}></canvas>;
+  return <canvas className='canvas' ref={canvas}></canvas>;
 }
