+++ packages/client/src/pages/LeaderBoard/styles.module.css
@@ -2,6 +2,7 @@
   height: calc(100vh - 80px);
   display: flex;
   flex-wrap: wrap;
+  overflow-y: auto;
 
   &__topLeader {
     display: flex;
@@ -11,7 +12,7 @@
     flex-basis: 50%;
     flex-grow: 1;
     min-width: 400px;
-    background-image: url('../../assets/imgs/leaderFon.webp');
+    background-image: url('src/assets/imgs/leaderFon.webp');
     background-repeat: no-repeat;
     background-size: cover;
   }
