+++ packages/client/src/pages/Profile/styles.module.css
@@ -1,8 +1,13 @@
+.wrapper {
+  width: 100vw;
+  height: 100vh;
+  overflow-y: auto;
+}
+
 .profile {
   display: flex;
   flex-direction: column;
   align-items: center;
-  overflow-y: auto;
   padding: 30px;
   margin: 25px auto;
   background: var(--color-normal);
@@ -24,6 +29,8 @@
 
   &__buttons {
     display: flex;
+    flex-wrap: wrap;
+    gap: 10px;
     align-items: center;
     justify-content: space-around;
     width: 100%;
